pyproject.toml
[project]
name = "de4"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiofiles==24.1.0",
    "annotated-types==0.7.0",
    "anyio==4.10.0",
    "bcrypt==4.3.0",
    "cachetools==5.5.2",
    "certifi==2025.8.3",
    "cffi==1.17.1",
    "charset-normalizer==3.4.3",
    "click==8.2.1",
    "colorama==0.4.6",
    "cryptography==45.0.7",
    "distro==1.9.0",
    "ecdsa==0.19.1",
    "fastapi==0.116.1",
    "filetype==1.2.0",
    "google-ai-generativelanguage==0.6.18",
    "google-api-core==2.25.1",
    "google-auth==2.40.3",
    "googleapis-common-protos==1.70.0",
    "grpcio==1.74.0",
    "grpcio-status==1.74.0",
    "h11==0.16.0",
    "httpcore==1.0.9",
    "httptools==0.6.4",
    "httpx==0.28.1",
    "idna==3.10",
    "jinja2==3.1.6",
    "jiter==0.10.0",
    "jsonpatch==1.33",
    "jsonpointer==3.0.0",
    "langchain-core==0.3.75",
    "langchain-google-genai==2.1.10",
    "langchain-openai==0.3.32",
    "langgraph==0.6.6",
    "langgraph-checkpoint==2.1.1",
    "langgraph-prebuilt==0.6.4",
    "langgraph-sdk==0.2.6",
    "langsmith==0.4.25",
    "markupsafe==3.0.2",
    "openai==1.106.1",
    "orjson==3.11.3",
    "ormsgpack==1.10.0",
    "packaging==25.0",
    "passlib==1.7.4",
    "proto-plus==1.26.1",
    "protobuf==6.32.0",
    "psycopg2>=2.9.10",
    "pyasn1==0.6.1",
    "pyasn1-modules==0.4.2",
    "pycparser==2.22",
    "pydantic==2.11.7",
    "pydantic-core==2.33.2",
    "python-dotenv==1.1.1",
    "python-jose==3.5.0",
    "python-multipart==0.0.20",
    "pyyaml==6.0.2",
    "regex==2025.9.1",
    "requests==2.32.5",
    "requests-toolbelt==1.0.0",
    "rsa==4.9.1",
    "six==1.17.0",
    "sniffio==1.3.1",
    "sqlalchemy>=2.0.43",
    "starlette==0.47.3",
    "tenacity==9.1.2",
    "tiktoken==0.11.0",
    "tqdm==4.67.1",
    "typing-extensions==4.15.0",
    "typing-inspection==0.4.1",
    "urllib3==2.5.0",
    "uvicorn==0.35.0",
    "watchfiles==1.1.0",
    "websockets==15.0.1",
    "xxhash==3.5.0",
    "zstandard==0.24.0",
]
